+++ src/constants.ts
@@ -1,15 +1,15 @@
 //core entities
 export const GITHUB_ACCOUNT_ENTITY_TYPE = 'github_account';
-export const GITHUB_ACCOUNT_ENTITY_CLASS = 'Account';
+export const GITHUB_ACCOUNT_ENTITY_CLASS = ['Account'];
 export const GITHUB_ACCOUNT_MEMBER_RELATIONSHIP_TYPE =
   'github_account_has_user';
 export const GITHUB_ACCOUNT_TEAM_RELATIONSHIP_TYPE = 'github_account_has_team';
 export const GITHUB_ACCOUNT_REPO_RELATIONSHIP_TYPE = 'github_account_owns_repo';
 
 export const GITHUB_MEMBER_ENTITY_TYPE = 'github_user';
-export const GITHUB_MEMBER_ENTITY_CLASS = 'User';
+export const GITHUB_MEMBER_ENTITY_CLASS = ['User'];
 export const GITHUB_COLLABORATOR_ENTITY_TYPE = 'github_user';
-export const GITHUB_COLLABORATOR_ENTITY_CLASS = 'User';
+export const GITHUB_COLLABORATOR_ENTITY_CLASS = ['User'];
 export const GITHUB_MEMBER_ACCOUNT_RELATIONSHIP_TYPE =
   'github_user_manages_account';
 export const GITHUB_MEMBER_TEAM_RELATIONSHIP_TYPE = 'github_user_manages_team';
@@ -21,21 +21,21 @@ export const GITHUB_MEMBER_OPENED_PR_RELATIONSHIP_TYPE =
   'github_user_opened_pullrequest';
 
 export const GITHUB_REPO_ENTITY_TYPE = 'github_repo';
-export const GITHUB_REPO_ENTITY_CLASS = 'CodeRepo';
+export const GITHUB_REPO_ENTITY_CLASS = ['CodeRepo'];
 export const GITHUB_REPO_PR_RELATIONSHIP_TYPE = 'github_repo_has_pullrequest';
 
 export const GITHUB_PR_ENTITY_TYPE = 'github_pullrequest';
-export const GITHUB_PR_ENTITY_CLASS = 'PR';
+export const GITHUB_PR_ENTITY_CLASS = ['PR'];
 
 export const GITHUB_TEAM_ENTITY_TYPE = 'github_team';
-export const GITHUB_TEAM_ENTITY_CLASS = 'UserGroup';
+export const GITHUB_TEAM_ENTITY_CLASS = ['UserGroup'];
 export const GITHUB_TEAM_MEMBER_RELATIONSHIP_TYPE = 'github_team_has_user';
 export const GITHUB_REPO_TEAM_RELATIONSHIP_TYPE = 'github_repo_allows_team';
 export const GITHUB_REPO_USER_RELATIONSHIP_TYPE = 'github_repo_allows_user';
 
 //apps
 export const GITHUB_APP_ENTITY_TYPE = 'github_app';
-export const GITHUB_APP_ENTITY_CLASS = 'Application';
+export const GITHUB_APP_ENTITY_CLASS = ['Application'];
 export const GITHUB_ACCOUNT_APP_RELATIONSHIP_TYPE =
   'github_account_installed_app';
 
@@ -46,7 +46,7 @@ export const GITHUB_ENV_SECRET_ENTITY_TYPE = 'github_env_secret';
 export const GITHUB_SECRET_ENTITY_CLASS = ['Secret'];
 export const GITHUB_ENVIRONMENT_ENTITY_TYPE = 'github_environment';
 export const GITHUB_ENVIRONMENT_ENTITY_CLASS = ['Configuration'];
-//org sercrets
+//org secrets
 export const GITHUB_ACCOUNT_SECRET_RELATIONSHIP_TYPE =
   'github_account_has_org_secret';
 export const GITHUB_REPO_ORG_SECRET_RELATIONSHIP_TYPE =
@@ -68,6 +68,15 @@ export const GITHUB_ENV_SECRET_ORG_SECRET_RELATIONSHIP_TYPE =
   'github_env_secret_overrides_org_secret';
 export const GITHUB_ENV_SECRET_REPO_SECRET_RELATIONSHIP_TYPE =
   'github_env_secret_overrides_repo_secret';
+
+//issues
+export const GITHUB_ISSUE_ENTITY_TYPE = 'github_issue';
+export const GITHUB_ISSUE_ENTITY_CLASS = ['Issue'];
+export const GITHUB_REPO_ISSUE_RELATIONSHIP_TYPE = 'github_repo_has_issue';
+export const GITHUB_MEMBER_CREATED_ISSUE_RELATIONSHIP_TYPE =
+  'github_user_created_issue';
+export const GITHUB_MEMBER_ASSIGNED_ISSUE_RELATIONSHIP_TYPE =
+  'github_user_assigned_issue';
 
 //these constants are names used to save and retrieve data between steps
 //they are constants instead of strings so that TypeScript will detect spelling errors